project: update configs to get project working
Hamza made some good changes to get rid of the redundancy in the reverse
proxy we had going from 9093 -> 50003. I had to disable the nginx
reverse proxy rules on the PI to get it working properly. I made a
backup of the reverse proxy rules for future reference. Also created
a gitmessage for future commits. Small fix in the
telemetry_frontend_pi_script

Signed-off-by: Your Name <issam.akhtar@gmail.com>